backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
pandas
supabase
python-dotenv
numpy